requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.24.3
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
TA-Lib==0.4.32
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
yarl==1.13.1
